requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorclass==2.2.2
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==5.0
django-cors-headers==4.3.1
django-filter==23.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
docopt==0.6.2
idna==3.6
oauthlib==3.2.2
packaging==23.2
pip-upgrader==1.4.15
psycopg==3.1.14
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
terminaltables==3.1.10
typing_extensions==4.8.0
urllib3==2.1.0
